requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2
requests==2.31.0
python-dotenv==1.0.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.45
langchain-huggingface
langchain-text-splitters==0.3.7
sentence-transformers==2.6.1
faiss-cpu
pypdf